Total Tank!
OK so this is the first group release of code. This is my version of what we did on Saturday, and is what I believe to be the first functioning set of code (that one of us has made, anyway). I've commented it out so you all should be able to tell whats happening even if you aren't experienced. Thanks