MAJOR CHANGES THIS TIME
this is totally different file than the one we used before. All previous versions are a modified 'getting started' file, and this is my version the Iterative robot template version. We may not be using the java version due to the stupid complexity of simulations. I think we should consider buying parts for a desktop that we can use for the robotics room and execute simulations and stuff but we'll talk about that later.